EIP-161 collect_touched_accounts(): remove computation.is_origin_computation when considering SELFDESTRUCT beneficiaries + refactor.
`computation.is_origin_computation` has been present since the file was first introduced by @pipermerriam:

https://github.com/ethereum/py-evm/commit/69a8385476983b4a060f307061ab258027b64eff#diff-28f7a9ef4182d19a95f4cccb1c073ef3

However, whether it's present or not has no effect on the tests passing.

(Possibly, this is due to the case not being covered in the tests...)

I can't find a place which explicitly states that beneficiary collection
should only be skipped if the error happened in the origin computation.

From the fact that having a `computation.is_origin_computation` in the
next if-clause (`if computation.msg.to != constants.CREATE_CONTRACT_ADDRESS`)
caused an error in a RevertPrecompiledTouchExactOOG test, it would seem
that upstream (geth v1.9.6) does not consider this a necessary condition.

Therefore, removing the condition here, too; and re-working the if-else
clauses to look like they did previously (moving the RIPEMD160 special-casing
as a sub-clause of the main clause).